+++ CHANGELOG.md
@@ -19,6 +19,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Timewalking Vendor` details for Warlords of Draenor and Legion don't depend on calendar event data anymore.
 - Calendar data look-up process refurbished.
 
+### Fixed
+
+- Not all `Addon Compartment` icons have been displayed; using now mostly inline icons instead.
+
 ## [0.16.4+100100] - 2023-07-07
 
 ### Added